% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compCalc.R
\name{compCalc}
\alias{compCalc}
\title{Composite Correlation}
\usage{
compCalc(trData, quantile, fullMean, up = "up")
}
\arguments{
\item{trData}{tree ring data - probably will appear something like trDat$...}

\item{quantile}{selection of quantile numerically}

\item{fullMean}{the full dataset of seasonal means from year 1..n}

\item{up}{deciding whether upper or lower quantile for calculations, default at up}
}
\value{
com.c based on choice in quantile.
}
\description{
This function allows you to create composite spatial field correlations
}
