% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrCL.R
\name{detrCL}
\alias{detrCL}
\title{Dating of netcdf Files and creating a seasonal compilation}
\usage{
detrCL(climData, mod)
}
\arguments{
\item{climData}{climate data}

\item{mod}{mode of detrending}
}
\value{
results from datM
}
\description{
This function allows you to create spatial field correlations with tree rings and
gridded climate data
}
