% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullCorr.R
\name{fullCorr}
\alias{fullCorr}
\title{Spatial Field Correlation Function}
\usage{
fullCorr(x, y)
}
\arguments{
\item{x}{climate data}

\item{y}{tree ring data}
}
\value{
results from temp
}
\description{
This function allows you to create spatial field correlations with tree rings and
gridded climate data
}
