% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdfRead.R
\name{ncdfRead}
\alias{ncdfRead}
\title{Reading in netCDF Files}
\usage{
ncdfRead(ncdf)
}
\arguments{
\item{ncdf}{tree ring data - probably will appear something like trDat$...}
}
\value{
dat based on choice in quantile.
}
\description{
This allows users to choose climate variables interactively and creates a raster brick
}
