% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasNm.R
\name{seasNm}
\alias{seasNm}
\title{Dating of netcdf Files and creating a seasonal compilation}
\usage{
seasNm(climDat, hem, lg, FUN)
}
\arguments{
\item{climDat}{climate data}

\item{hem}{hemisphere of tree ring data}

\item{lg}{lag of seasonal effect}

\item{FUN}{function to be used in stackApply}
}
\value{
results from datM
}
\description{
This function allows you to create spatial field correlations with tree rings and
gridded climate data
}
